Reordered files to have a more concise structure
