+++ src/idls/dab_registries/canister_registry.did.ts
