+++ src/idls/dab_registries/registry_standard.did.ts
