+++ src/index.ts
@@ -1,10 +1,10 @@
 import { HttpAgent } from '@dfinity/agent';
 import { Principal } from '@dfinity/principal';
 import { NFTCollection, NFTStandards, NFTDetails } from './interfaces/nft';
-import NFT from './nft_standards/default';
-import EXT from './nft_standards/ext';
-import ICPunks from './nft_standards/ic_punks';
+import NFT from './standard_wrappers/nft_standards/default';
+import EXT from './standard_wrappers/nft_standards/ext';
+import ICPunks from './standard_wrappers/nft_standards/ic_punks';
 
 export * from './interfaces/nft';
-export * from './dab_canisters';
+export * from './registries';
 export { default as standards } from './constants/standards';
