+++ src/interfaces/dab_registries/canister_registry.ts
