+++ src/interfaces/dab_registries/registry_standard.ts
