+++ src/interfaces/nft.ts
@@ -1,7 +1,7 @@
-import DepartureLabs from '../nft_standards/departure_labs';
-import DIP721 from '../nft_standards/dip_721';
-import EXT from '../nft_standards/ext';
-import ICPunks from '../nft_standards/ic_punks';
+import DepartureLabs from '../standard_wrappers/nft_standards/departure_labs';
+import DIP721 from '../standard_wrappers/nft_standards/dip_721';
+import EXT from '../standard_wrappers/nft_standards/ext';
+import ICPunks from '../standard_wrappers/nft_standards/ic_punks';
 
 export type NFTStandards = typeof EXT | typeof ICPunks | typeof DepartureLabs | typeof DIP721;
