+++ src/registries/index.ts
