+++ src/standard_wrappers/nft_standards/ic_punks.ts
@@ -1,12 +1,12 @@
 import { Actor, ActorSubclass, HttpAgent } from '@dfinity/agent';
 import { Principal } from '@dfinity/principal';
 
-import NFT_ICPUNKS, { TokenDesc } from '../interfaces/icpunks';
-import IDL from '../idls/icpunks.did';
+import NFT_ICPUNKS, { TokenDesc } from '../../interfaces/icpunks';
+import IDL from '../../idls/icpunks.did';
 import NFT from './default';
-import { NFTDetails } from '../interfaces/nft';
-import { NFT_CANISTERS } from '../constants/canisters';
-import { NFT as NFTStandard} from '../constants/standards';
+import { NFTDetails } from '../../interfaces/nft';
+import { NFT_CANISTERS } from '../../constants/canisters';
+import { NFT as NFTStandard} from '../../constants/standards';
 
 const getICPBunnyCanisterId = (index) =>
   NFT_CANISTERS.ICP_BUNNY_STORAGE[index % 10];
