+++ src/standard_wrappers/token_standards/icpMethods.ts
@@ -2,8 +2,8 @@
 import { Principal } from '@dfinity/principal';
 import { ActorSubclass } from '@dfinity/agent';
 
-import LedgerService from '../interfaces/ledger';
-import { Metadata } from '../interfaces/ext';
+import LedgerService from '../../interfaces/ledger';
+import { Metadata } from '../../interfaces/ext';
 import {
   Balance,
   BurnParams,
@@ -12,8 +12,8 @@ import {
   SendParams,
   SendResponse,
 } from './methods';
-import { BaseMethodsExtendedActor } from '../utils/actorFactory';
-import { getAccountId } from '../utils/account';
+import { BaseMethodsExtendedActor } from '../../utils/actorFactory';
+import { getAccountId } from '../../utils/account';
 
 type BaseLedgerService = BaseMethodsExtendedActor<LedgerService>;
