+++ src/standard_wrappers/token_standards/methods.ts
@@ -1,9 +1,9 @@
 import { ActorSubclass } from '@dfinity/agent';
 import { Principal } from '@dfinity/principal';
 
-import { Metadata } from '../interfaces/ext';
-import { BurnResult } from '../interfaces/xtc';
-import { BaseMethodsExtendedActor } from '../utils/actorFactory';
+import { Metadata } from '../../interfaces/ext';
+import { BurnResult } from '../../interfaces/xtc';
+import { BaseMethodsExtendedActor } from '../../utils/actorFactory';
 
 export type SendResponse =
   | { height: string }
